pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["witelokk <witelokk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.29.0"
fastapi = "^0.110.2"
sqlalchemy = "^2.0.29"
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-multipart = "^0.0.9"
pyjwt = "^2.8.0"
pydantic-settings = "^2.2.1"
reportlab = "^4.2.0"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
